requirements.txt
discord.py
Flask
aiohttp
aiofiles
gunicorn
boto3
awscli
requests
jsonschema
pyyaml
orjson